Update core.py
urllib.pathname2url will fail when the to_translate variable contains one or more ":" (=colon). It will raise a IOError. Because this function is meant to use for urls and the colon gets handled in a specific way because of windows path's. 
Since to_translate is not really a path i propose using urllib.quote_plus.
When using urllib.quote_plus the error will not raise, it will only quote the to_translate and replacing " " (=space) with "+" (=plus) and nothing more.